% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{variable name}
\name{A.t}
\alias{A.t}
\title{A.t}
\format{8-column matrix, each row correspond to a local-ATAC-QTL}
\description{
trinarized local ATAC-seq signal
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
