% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis.R
\name{as.data.frame.infima_results}
\alias{as.data.frame.infima_results}
\title{as.data.frame for infima_results}
\usage{
\method{as.data.frame}{infima_results}(x, n_cores = detectCores() - 2, ...)
}
\arguments{
\item{x}{The S3 class \code{infima_results}}

\item{n_cores}{The number of cores}

\item{...}{Additional arguments}
}
\value{
a data frame containing the following colums:
\tabular{ll}{
\code{snp_id ... alt} \tab SNP information of the causal local-ATAC-QTLs \cr
\code{ensembl ... lod} \tab DO-eQTL data: DO gene & marker information \cr
\code{footprint} \tab the footprint information \cr
\code{footprint.rs} \tab the footprint information rank score \cr
\code{dist} \tab the distance score component in the prior \cr
\code{dist.rs} \tab the distance rank score \cr
\code{cor.A.E} \tab the Correlation between ATAC-seq signal and gene expression \cr
\code{cor.A.E.rs} \tab the cor.A.E rank score \cr
\code{cor.A.B} \tab the Correlation between ATAC-seq signal and founder effect size \cr
\code{cor.A.B.rs} \tab the cor.A.B rank score \cr
\code{p} \tab the total number of candidates \cr
\code{k} \tab the number of candidates in the credible set \cr
\code{Z} \tab the posterior probabilities of the output \cr
\code{Z.rs} \tab the rank score of the posterior probabilities \cr
}
}
\description{
The S3 method for the class \code{infima_results}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
