% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{variable name}
\name{B.t}
\alias{B.t}
\title{B.t}
\format{a vector}
\description{
trinarized founder gene expressions (averaged over strains)
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
