% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pseudocount.R
\name{compute_prior}
\alias{compute_prior}
\title{Compute Dirichlet prior}
\usage{
compute_prior(
  raw_data = NULL,
  model_data = NULL,
  n_cores = detectCores() - 2,
  verbose = TRUE
)
}
\arguments{
\item{raw_data}{The output from \code{raw_input_data}}

\item{model_data}{The output from  \code{model_input_data}}

\item{n_cores}{The number of cores for parallel computing.}

\item{verbose}{Print messages or not. 
Default = \code{detectCores() - 2}.}
}
\value{
A prior data object.

\tabular{ll}{
\code{PI} \tab The overall prior term. \cr
\code{dist} \tab The distance score. \cr
\code{F.g} \tab The footprint annotation. \cr
\code{cor.A.B} \tab The Correlation betwwen ATAC-seq signal and founder effect size. \cr
\code{cor.A.E} \tab The Correlation between ATAC-seq signal and gene expression. \cr
}
}
\description{
For the candidate SNPs of each row of DO-eQTL data, 
keep track of their the distance scores, footprinting annotations,
correlations between local ATAC-seq signal and effect size,
correlations between local ATAC-seq signal and gene expression.
}
\examples{
data('example-10-genes')
raw_data <- raw_input_data(dt1, dt2, dt3)
model_data <- model_input_data(raw_data)
prior <- compute_prior(raw_data, model_data)
}
\seealso{
\code{\link{raw_input_data}}, \code{\link{model_input_data}}.
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
