% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pseudocount.R
\name{compute_pseudocount}
\alias{compute_pseudocount}
\title{Compute pseudocount prior}
\usage{
compute_pseudocount(raw_data = NULL, model_data = NULL, window = 1e+06)
}
\arguments{
\item{raw_data}{The output from @seealso raw_input_data}

\item{model_data}{The output from @seealso model_input_data}

\item{window}{The window around DO-eQTL marker contains candidate SNPs.
Default = 1 Mbp}
}
\value{
A list of data objects.
PI: The pseudocount term.
dist: The distance score.
F.g: The footprint information.
cor.A.R: correlation betwwen ATAC-seq signal and effect size.
cor.A.E: correlation between ATAC-seq signal and gene expression.
}
\description{
Compute pseudocount prior
}
