% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{cor.A.R}
\alias{cor.A.R}
\title{cor.A.R}
\format{a vector, the dimension of which equals to the number of local-ATAC-QTLs}
\description{
the correlation between ATAC-seq signal and founder effect size
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
