% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{data_trinarize}
\alias{data_trinarize}
\title{Data matrix trinarization across 8 founder strains.}
\usage{
data_trinarize(mat, cutoff = 0.2)
}
\arguments{
\item{mat}{The numeric matrix with 8 columns.}

\item{cutoff}{A value between 0 and 0.5. Default 0.2.}
}
\value{
A numeric matrix after trinarization.
}
\description{
This function trinarizes numeric matrix rowwisely
to \{-1,0,1\}. The matrix should have 8 columns corresponding to
8 founder strains \code{129, AJ, B6, CAST, NOD, NZO, PWK, WSB}.

First, each row of the matrix is standardized to [0,1].

Then, all values subtracted by the \code{B6} value in order to
set \code{B6} as the reference.

Finally, anything larger than \code{cutoff} are transformed to 1.
Anything smaller than -\code{cutoff} are transformed to -1.
Anything between are transformed to 0.
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
