% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{variable name}
\name{dist}
\alias{dist}
\title{dist}
\format{a vector, the dimension of which equals to the number of local-ATAC-QTLs}
\description{
the distance score component in the prior
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
