% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{variable name}
\name{E.t}
\alias{E.t}
\title{E.t}
\format{8-column matrix, each row correspond to a local-ATAC-QTL}
\description{
the trinarized founder allelic effects
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
