% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{variable name}
\name{footprint.rs}
\alias{footprint.rs}
\title{footprint.rs}
\format{a vector, the dimension of which equals to the number of local-ATAC-QTLs}
\description{
the footprint rank score vector
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
