% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{infima}
\alias{infima}
\title{infima data object}
\format{A list object.}
\description{
An infima object containing the results after model fitting.

Parameters (see the paper for details):
\tabular{ll}{
\code{alpha} \tab The estimated \code{a1}. \cr
\code{beta} \tab The estimated \code{b1}. \cr
\code{alpha.0} \tab The estimated \code{a0}. \cr
\code{beta.0} \tab The estimated \code{b0}. \cr
\code{gamma} \tab The estimated \code{gamma}. \cr
}
The results of INFIMA model:
\tabular{ll}{
\code{em.iter} \tab The number of iterations used. \cr
\code{V.g} \tab The posterior probability for each DO gene to be causal. \cr
\code{Z.g} \tab The posterior probability of candidate SNPs for each DO gene. \cr
}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
