% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{infima_results}
\alias{infima_results}
\title{infima_results data object}
\format{A list object.}
\description{
\code{infima_results$infima} is an \code{infima} object.
\code{infima_results$results} is a list with length equal to the total number
of instances in the DO-eQTL data. \code{infima_results$results[[i]]} is \code{NULL} if the i-th DO gene does
not pass the FDR cutoff. Otherwise, \code{infima_results$results[[i]]} contains two lists \code{input_data}
and \code{output_data}.

\code{input_data} contains:
\tabular{ll}{
\code{Y} \tab raw DO-eQTL signal \cr
\code{Y.t} \tab trinarized DO-eQTL signal \cr
\code{A} \tab raw ATAC-seq signal \cr
\code{A.t} \tab trinarized ATAC-seq signal \cr
\code{B} \tab raw RNA-seq data (all samples) \cr
\code{B.avg} \tab raw RNA-seq data (averaged) \cr
\code{B.t} \tab trinarized RNA-seq data \cr
\code{do.eqtl} \tab DO-eQTL data: DO gene & marker information \cr
\code{snpData} \tab SNP information of causal local-ATAC-MVs \cr
\code{D} \tab the edit distance \cr
\code{footprint} \tab the footprint information \cr
\code{E.t} \tab the trinarized allelic effects \cr
\code{dist} \tab the distance score component in the prior \cr
\code{cor.A.E} \tab the Correlation between ATAC-seq signal and founder genotype effect \cr
\code{cor.A.B} \tab the Correlation between ATAC-seq signal and gene expression \cr
\code{p} \tab the total number of candidates \cr
\code{k} \tab the number of candidates in the credible set \cr
}

\code{output_data} contains:
\tabular{ll}{
\code{footprint.rs} \tab the footprint information rank score \cr
\code{dist.rs} \tab the distance rank score \cr
\code{cor.A.E.rs} \tab the cor.A.E rank score \cr
\code{cor.A.B.rs} \tab the cor.A.B rank score \cr
\code{Z} \tab the posterior probabilities of the output \cr
\code{Z.rs} \tab the rank score of the posterior probabilities \cr
}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
