% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{input_query}
\alias{input_query}
\title{input_query data object}
\format{A list object.}
\description{
A list of multi-omics input data correspond to the query (snp_id, ensembl, qtl_marker).
\tabular{ll}{
  \code{Y} \tab DO allelic effects \cr
  \code{Y.t} \tab DO allelic effects (trinarized, \code{B6} as reference) \cr
  \code{A} \tab local ATAC-seq signal \cr
  \code{A.t} \tab local ATAC-seq signal (trinarized, \code{B6} as reference) \cr
  \code{B} \tab founder gene expressions from all samples \cr
  \code{B.avg} \tab average founder gene expressions from all strains \cr
  \code{B.t} \tab average founder gene expressions from all strains (trinarized, \code{B6} as reference) \cr
  \code{D} \tab edit distance, the difference between Y.t and E.t \cr
  \code{E.t} \tab founder genotype effects (trinarized, \code{B6} as reference) \cr
  \code{snpData} \tab the information of the query local-ATAC-MV \cr
  \code{snp_id} \tab the SNP id of the query local-ATAC-MV \cr
  \code{ensembl} \tab the ensembl id of the query DO gene \cr
  \code{qtl_marker} \tab the name of the QTL marker of the query DO gene, a DO gene may correspond to multiple markers \cr
}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
