% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{model_data}
\alias{model_data}
\title{model_data data object}
\format{A list object.}
\description{
The processed input data.
\tabular{ll}{
\code{E.g} \tab Trinarized effect size of candidate SNPs with respect to DO gene. \cr
\code{Y.g} \tab Trinarized DO-eQTL allelic dependence. \cr
\code{A.g} \tab Trinarized local ATAC-seq signal for candidate SNPs. \cr
\code{F.g} \tab Footprint annotations for candidate SNPs. \cr
\code{D.g} \tab Edited distances between candidate SNP and DO-eQTL signal. \cr
\code{B.g} \tab Trinarized gene expression in founder RNA-seq data. \cr
\code{snp_index} \tab The indices of candidate SNPs in the original list of SNPs. \cr
\code{p.g} \tab The number of candidate SNPs for each DO gene g. \cr
\code{window} \tab The window size in Mb. \cr
}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
