% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{model_fitting}
\alias{model_fitting}
\title{INFIMA model fitting}
\usage{
model_fitting(
  model_data = NULL,
  prior = NULL,
  alpha_init = c(0.6, 0.3, 0.1),
  beta_init = c(0.5, 0.4, 0.1),
  gamma_init = 0.5,
  max_iter = 100,
  verbose = TRUE
)
}
\arguments{
\item{model_data}{The output from \code{\link{model_input_data}}.}

\item{prior}{The output from \code{\link{compute_prior}}.}

\item{alpha_init}{The initial parameters of alpha:
The three-dimensional multinomial parameters for the generative model 
of normal strains \code{129, AJ, B6, NOD, NZO}.
Default \code{alpha_init = c(0.6, 0.3, 0.1)}.}

\item{beta_init}{The initial parameters of beta:
The three-dimensional multinomial parameters for the generative model 
of wild strains \code{CAST, PWK, WSB}.
Default \code{alpha_init = c(0.5, 0.4, 0.1)}.}

\item{gamma_init}{The initial value of gamma: 
the grand probability of a DO gene to be causal.
Default \code{gamma_init = 0.5}.}

\item{max_iter}{The maximum iteration allowed. 
Default \code{max_iter = 100}.}

\item{verbose}{Print messages or not. 
Default \code{verbose = TRUE}.}
}
\value{
An infima object containing the results after model fitting.

Parameters (see the paper for details):
\tabular{ll}{
\code{alpha} \tab The estimated \code{a1}. \cr
\code{beta} \tab The estimated \code{b1}. \cr
\code{alpha.0} \tab The estimated \code{a0}. \cr
\code{beta.0} \tab The estimated \code{b0}. \cr
\code{gamma} \tab The estimated \code{gamma}. \cr
}
The results of INFIMA model:
\tabular{ll}{
\code{em.iter} \tab The number of iterations used. \cr
\code{V.g} \tab The posterior probability for each DO gene to be causal. \cr
\code{Z.g} \tab The posterior probability of candidate SNPs for each DO gene. \cr
}
}
\description{
The main function for INFIMA model fitting.
Taking inputs from founder ATAC-seq, founder RNA-seq, DO-eQTL
as well as the footprinting annotation data, INFIMA outputs SNP-level
quantifications to pinpoint the causal variants.
}
\examples{
data('example-10-genes')
raw_data <- raw_input_data(dt1, dt2, dt3)
model_data <- model_input_data(raw_data)
prior <- compute_prior(raw_data, model_data)
infima <- model_fitting(model_data, prior)

}
\seealso{
\code{\link{raw_input_data}}, \code{\link{model_input_data}}, \code{\link{compute_prior}}.
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
