% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{model_input_data}
\alias{model_input_data}
\title{Process the raw input data}
\usage{
model_input_data(
  raw_data = NULL,
  window = 1e+06,
  cutoff = 0.2,
  n_cores = detectCores() - 2
)
}
\arguments{
\item{raw_data}{The output from @seealso raw_input_data}

\item{window}{The window around DO-eQTL marker contains candidate SNPs.
Default = 1 Mbp}

\item{cutoff}{Trinarization cutoff
Default = 0.2}
}
\value{
A list of processed data objects
E.g: Trinarized effect size of candidate SNPs with respect to DO gene
Y.g: Trinarized DO-eQTL allelic dependence
A.g: Trinarized local ATAC-seq signal for candidate SNPs
F.g: Footprint annotations for candidate SNPs
D.g: Edited distances between candidate SNP and DO-eQTL signal
B.g: Trinarized gene expression in founder RNA-seq data
snp_index: The indices of candidate SNPs in the original list of SNPs
p.g: The number of candidate SNPs for each DO gene g.
}
\description{
Process the raw input data
}
\examples{
model_data <- model_input_data(raw_data)
}
