% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis.R
\name{plot}
\alias{plot}
\alias{plot.input_query}
\title{plot of input_query}
\usage{
\method{plot}{input_query}(x, option = NULL)
}
\arguments{
\item{x}{The \code{input_query} S3 class}

\item{option}{Which piece of data to plot
1: DO allele effect data
2: ATAC-seq data
3: founder gene expression data
4: edit distance and founder allele efect}

\item{...}{Additional arguments}
}
\description{
S3 method for the class \code{input_query}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
