% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis.R
\name{plot_input}
\alias{plot_input}
\title{plot the query input data}
\usage{
plot_input(x, option = NULL, ...)
}
\arguments{
\item{x}{The \code{input_query} S3 class}

\item{option}{Which piece of data to plot

1: DO allele effect data

2: ATAC-seq data

3: founder gene expression data

4: edit distance and founder allele efect}

\item{...}{Additional arguments}
}
\value{
Plots
}
\description{
S3 method for the class \code{input_query}
}
\examples{
plot_input(input_query, option = 1) ## DO allele effect data
plot_input(input_query, option = 2) ## ATAC-seq data
plot_input(input_query, option = 3) ## RNA-seq data
plot_input(input_query, option = 4) ## edit distance and founder allele effect


}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
