% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{prior}
\alias{prior}
\title{prior data object}
\format{A list object.}
\description{
The Dirichlet prior information

\tabular{ll}{
\code{PI} \tab The overall prior term. \cr
\code{dist} \tab The distance score. \cr
\code{F.g} \tab The footprint annotation. \cr
\code{cor.A.B} \tab The Correlation betwwen ATAC-seq signal and founder effect size. \cr
\code{cor.A.E} \tab The Correlation between ATAC-seq signal and gene expression. \cr
}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
