% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis.R
\name{query_input_data}
\alias{query_input_data}
\title{Given query snp_id and ensembl id, return the multi-omics input data.}
\usage{
query_input_data(x, snp_id = NULL, ensembl = NULL, qtl_marker = NULL, ...)
}
\arguments{
\item{x}{The S3 class \code{infima_results}}

\item{snp_id}{The query snp_id of local-ATAC-QTL}

\item{ensembl}{The query ensembl column value of DO-eQTL data}

\item{qtl_marker}{The query qtl_marker column value of DO-eQTL data}
}
\value{
an input_query class, which is a list of input data corresponding to the query
}
\description{
The S3 method for the class \code{infima_results}
}
\examples{
input_query <- query_input_data(infima_results, snp_id = 'rs51076312', ensembl = 'ENSMUSG00000037995', qtl_marker = '1_172713578')

}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
