% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{raw_data data object}
\format{A list object.}
\description{
The organized raw input data.
\tabular{ll}{
\code{YY} \tab DO-eQTL allelic dependence matrix. \cr
\code{AA} \tab Local ATAC-seq signal matrix. \cr
\code{EE} \tab Local-ATAC-QTL genotypes. \cr
\code{rna.seq} \tab Full founder RNA-seq data matrix. \cr
\code{BB} \tab Averaged founder RNA-seq data matrix. \cr
\code{FF} \tab Footprint annnotation vector. \cr
\code{do.eqtl} \tab DO-eQTL marker and target DO gene information. \cr
\code{snpData} \tab SNP information. \cr
}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
