% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{raw_input_data}
\alias{raw_input_data}
\title{Load INFIMA inputs from files.}
\usage{
raw_input_data(filename1 = NULL, filename2 = NULL, filename3 = NULL, ...)
}
\arguments{
\item{filename1}{DO-eQTL data, a data frame object or
the path to a comma separated .csv file.
Must contain the following columns with exactly the following column names in order:
\tabular{ll}{
ensembl \tab The Ensemble ID of DO-eQTL gene.\cr
symbol \tab The symbol of DO-eQTL gene.\cr
chr \tab The chromosome of DO-eQTL gene.\cr
start \tab The start position of DO-eQTL gene.\cr
end \tab The end position of DO-eQTL gene.\cr
strand \tab The strand of DO-eQTL gene.\cr
qtl_marker \tab The user-defined ID of DO-eQTL marker.\cr
qtl_chr \tab The chromosome of DO-eQTL marker.\cr
qtl_pos \tab The position of DO-eQTL marker.\cr
lod \tab The LOD score of DO-eQTL marker.\cr
}
Plus 8 columns named \code{129, AJ, B6, Cast, NOD, NZO, PWK, WSB}
recording the DO-eQTL allelic dependence measures for the 8 strains.}

\item{filename2}{Founder RNA-seq data, a matrix, a data frame or
the path to a comma separated .csv file.
DO gene (row) by sample (column) matrix with normalized gene expressions.
The column names must be strain names \code{129, AJ, B6, Cast, NOD, NZO, PWK, WSB}
and there must be at least one sample for each strain.
One caveat is that the rows of founder RNA-seq data should
correspond to the rows of DO-eQTL data (the same DO gene), i.e. the first two input files
have the same number of rows.}

\item{filename3}{SNP and local ATAC-seq signal data, a data frame object
or the path to a comma separated .csv file.
Must contain the following columns
with exactly the following column names in order:
\tabular{ll}{
snp_id \tab The SNP ID. \cr
chr \tab The chromosome of the SNP. \cr
snp_pos \tab The position of the SNP. \cr
ref \tab The reference nucleotide. \cr
alt \tab The alternative nucleotide. \cr
footprint \tab The user-defined footprint annotation.
The recommended range is between 0 and 1. The larger the value,
the more confidence the SNP is affecting a footprint. Put 0 if
the footprint analysis is not applicable. \cr
}
Plus

(a) 8 columns named as \code{strain-genotype} where \code{strain}
is \code{129, AJ, B6, Cast, NOD, NZO, PWK, WSB} in order. 0 means reference
nucleotide while 2 means alternative nucleotide.

(b) 8 columns named as \code{strain-ATAC-seq} where \code{strain}
is \code{129, AJ, B6, Cast, NOD, NZO, PWK, WSB} in order. The normalized
local ATAC-seq signal of the SNP.}
}
\value{
A raw_data object including:
\tabular{ll}{
\code{YY} \tab DO-eQTL allelic dependence matrix. \cr
\code{AA} \tab Local ATAC-seq signal matrix. \cr
\code{EE} \tab Local-ATAC-QTL genotypes. \cr
\code{rna.seq} \tab Full founder RNA-seq data matrix. \cr
\code{BB} \tab Averaged founder RNA-seq data matrix. \cr
\code{FF} \tab Footprint annnotation vector. \cr
\code{do.eqtl} \tab DO-eQTL marker and target DO gene information. \cr
\code{snpData} \tab SNP information. \cr
}
}
\description{
This function reads in the processed DO-eQTL, ATAC-seq
and RNA-seq data as inputs for INFIMA.
}
\examples{
data('example-10-genes')
raw_data <- raw_input_data(dt1, dt2, dt3)
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
