% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis.R
\name{snp_link_gene}
\alias{snp_link_gene}
\title{Link SNPs to effector genes}
\usage{
snp_link_gene(infima, model_data, fdr = 0.1, cum.pprob = 0.8, cred.set = 0.5)
}
\arguments{
\item{infima}{The output from \code{model_fitting}}

\item{model_data}{The output from \code{\link{model_input_data}}.}

\item{fdr}{The False Discovery Rate cutoff with direct posterior probability
approach (Newton, 2004). Default 0.1.}

\item{cum.pprob}{The maximum cumulative posterior probability for
the target credible set. Default 0.8. Take values between 0 and 1.}

\item{cred.set}{The maximum proportion of candidate SNPs allowed
in the credible set. Default 0.5. Take values between 0 and 1.}
}
\value{
A data table with the following columns.
\tabular{ll}{
\code{snp_id} \tab The SNP ID. \cr
\code{chr} \tab The chromosome of the SNP. \cr
\code{snp_pos} \tab The locus of the SNP. \cr
\code{ref} \tab The reference allele of the SNP. \cr
\code{alt} \tab The alternative allele of the SNP. \cr
\code{ensembl} \tab The ensemble ID of the effector gene. \cr
\code{symbol} \tab The gene symbol of the effector gene. \cr
}
}
\description{
Analyze the outputs from INFIMA and link local-ATAC-QTLs to effector genes
}
\examples{
data('example-10-genes')
raw_data <- raw_input_data(dt1, dt2, dt3)
model_data <- model_input_data(raw_data)
pseudocount <- compute_pseudocount(raw_data, model_data)
infima <- model_fitting(model_data, pseudocount)
effector_genes <- snp_link_gene(infima, model_data, fdr = 0.05, cum.pprob = 0.8, cred.set = 0.5)
}
\seealso{
\code{\link{model_fitting}}, \code{\link{model_input_data}}.
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
