% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{variable name}
\name{snpData}
\alias{snpData}
\title{snpData}
\format{a data table}
\description{
SNP information of causal local-ATAC-QTLs
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
