% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis.R
\name{summary}
\alias{summary}
\alias{summary.infima_results}
\alias{summary.infima}
\title{summary}
\usage{
summary(x)

\method{summary}{infima_results}(x, ...)

\method{summary}{infima}(x, ...)
}
\arguments{
\item{x}{The \code{infima} S3 class}

\item{...}{Additional arguments}
}
\description{
S3 method for the class \code{infima_results}

S3 method for the class \code{infima}
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}

Chenyang Dong \email{cdong@stat.wisc.edu}
}
