% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{variable name}
\name{Z.rs}
\alias{Z.rs}
\title{Z.rs}
\format{a vector, the dimension of which equals to the number of local-ATAC-QTLs}
\description{
the rank score of the posterior probabilities of the output
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
