% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{footprint}
\alias{footprint}
\title{footprint}
\format{a vector, the dimension of which equals to the number of local-ATAC-MVs}
\description{
the footprint vector
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
