% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{model_fitting}
\alias{model_fitting}
\title{INFIMA model fitting}
\usage{
model_fitting(
  model_data = NULL,
  pseudocount = NULL,
  alpha_init = c(0.6, 0.3, 0.1),
  beta_init = c(0.5, 0.4, 0.1),
  gamma_init = 0.5,
  max_iter = 100,
  verbose = TRUE
)
}
\arguments{
\item{model_data}{The output from @seealso model_input_data}

\item{pseudocount}{The output from @seealso compute_pseudocount}

\item{alpha_init}{The initial parameters of alpha:
The multinomial parameters for the generative model of normal strains.}

\item{beta_init}{The initial parameters of beta:
The multinomial parameters for the generative model of wild strains.}

\item{gamma_init}{The initial value of gamma: grand probability of causal gene.}

\item{max_iter}{The maximum iteration.}

\item{verbose}{Print outputs or not.}
}
\value{
A list of results after model fitting.
alpha: the estimated a1
beta: the estimated b1
alpha.0: the estimated a0
beta.0: the estimated b0
gamma: the estimated gamma
em.iter: the number of iterations used
V.g: the posterior probability for each DO gene to be causal
Z.g: the posterior probability of candidate SNPs for each DO gene
}
\description{
INFIMA model fitting
}
\examples{
infima <- model_fitting(model_data, pseudocount)
}
