% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{model_input_data}
\alias{model_input_data}
\title{Process the INFIMA inputs for each DO gene}
\usage{
model_input_data(
  raw_data = NULL,
  window = 1,
  cutoff = 0.2,
  n_cores = detectCores() - 2,
  verbose = TRUE
)
}
\arguments{
\item{raw_data}{The output from \code{\link{raw_input_data}}.}

\item{window}{The window size (Mbp) around DO-eQTL marker contains candidate SNPs.
Default = 1 Mbp.}

\item{cutoff}{Trinarization cutoff. A value between 0 and 0.5.
Default 0.2. See \code{\link{data_trinarize}}.}

\item{n_cores}{The number of cores for parallel computing.
Default = \code{detectCores() - 2}.}

\item{verbose}{Print outputs or not.
Default \code{verbose = TRUE}.}
}
\value{
A list of lists: the processed data objects are broken down
into small elements for each row of the DO-eQTL data.

\tabular{ll}{
\code{E.g} \tab Trinarized effect size of candidate SNPs with respect to DO gene. \cr
\code{Y.g} \tab Trinarized DO-eQTL allelic dependence. \cr
\code{A.g} \tab Trinarized local ATAC-seq signal for candidate SNPs. \cr
\code{F.g} \tab Footprint annotations for candidate SNPs. \cr
\code{D.g} \tab Edited distances between candidate SNP and DO-eQTL signal. \cr
\code{B.g} \tab Trinarized gene expression in founder RNA-seq data. \cr
\code{snp_index} \tab The indices of candidate SNPs in the original list of SNPs. \cr
\code{p.g} \tab The number of candidate SNPs for each DO gene g. \cr
\code{window} \tab The window size in Mbp. \cr
}
}
\description{
This function converts the raw input data into the data
for each row of the DO-eQTL data. For each eQTL marker, candidate SNPs
are searched from raw inputs within a user defined window size.
}
\examples{
data('example-10-genes')
raw_data <- raw_input_data(dt1, dt2, dt3)
model_data <- model_input_data(raw_data)
}
\seealso{
\code{\link{raw_input_data}}, \code{\link{data_trinarize}}.
}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
